tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/044.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: DOM mutation events when adding scripts: DOMNodeInsertedIntoDocument </title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

	<script>log('inline script #1');
		testlib.addScript('log(\'head script #1\')', {}, document.getElementsByTagName('head')[0], false, function(s){s.addEventListener( 'DOMNodeInsertedIntoDocument', function(){ log('DOMNodeInsertedIntoDocument'); }, false ); } );
		log('end script #1');
	</script>

	<script type="text/javascript">
	log( 'inline script #2' );
	var t = async_test()

	function test() {
		assert_array_equals(eventOrder, ['inline script #1', 'head script #1',  'DOMNodeInsertedIntoDocument', 'end script #1', 'inline script #2']);
		t.done();
}
	onload = t.step_func(test)
	</script>

</body></html>
